requirements.txt
appdirs==1.4.3
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
cycler==0.10.0
decorator==4.3.2
future==0.17.1
gmusicapi==12.0.0
gpsoauth==0.4.1
httplib2==0.12.1
idna==2.8
kiwisolver==1.0.1
lxml==4.3.1
matplotlib==3.0.2
MechanicalSoup==0.11.0
mock==2.0.0
mutagen==1.42.0
numpy==1.15.4
oauth2client==4.1.3
pbr==5.1.2
Pillow==5.4.1
proboscis==1.2.6.0
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycryptodomex==3.7.3
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
soupsieve==1.8
urllib3==1.24.1
validictory==1.1.2
wincertstore==0.2
